+++ src/lib.rs
@@ -1013,6 +1013,10 @@ fn main(mut cx: ModuleContext) -> NeonResult<()> {
         ctx::measureText,
     )?;
     cx.export_function(
+        "CanvasRenderingContext2D_textMetricsFields",
+        crate::node::typography::textMetricsFields,
+    )?;
+    cx.export_function(
         "CanvasRenderingContext2D_outlineText",
         ctx::outlineText,
     )?;